refactor: move funding program constants to utilities
- Updated import paths for funding program validation functions to use the new utilities directory.
- Removed the now obsolete funding-programs.ts file, centralizing funding program logic for better organization and maintainability.

These changes improve code structure and enhance clarity in funding program detection across the application.
